requirements.txt

langchain==0.3.26
langchain-community==0.3.26
langchain-openai==0.3.24
langchain-pinecone==0.2.8
flask==3.1.0
sentence-transformers==2.7.0
pypdf==5.1.0
python-dotenv==1.0.1
-e .
